fix logged neo velocity and add shooter feedforward

+++ robot/src/main/kotlin/org/team9432/lib/logged/neo/LoggedNeoIO.kt
@@ -37,7 +37,7 @@ interface LoggedNeoIO {
     fun setAngle(angle: Rotation2d) {}
 
     /* Run closed loop velocity control */
-    fun setSpeed(rpm: Int) {}
+    fun setSpeed(rpm: Double) {}
 
     fun setPID(p: Double, i: Double, d: Double) {}
